GA import does not choke if adgid is not specified on a landing page, and does not choke if article cannot be resolved

+++ server/convapi/aggregate.py
@@ -146,6 +146,7 @@ bounced_visits=%s, pageviews=%s, duration=%s',
 
         if data:
             dt, _, ad_group_id, _ = data.keys()[0]
+            assert(ad_group_id is not None)
 
             for stats in reports.models.ArticleStats.objects.filter(datetime=dt, ad_group=ad_group_id):
                 stats.reset_postclick_metrics()
@@ -214,6 +215,10 @@ bounced_visits=%s, pageviews=%s, duration=%s',
 
         ad_group_id = LandingPageUrl(entries[0]['Landing Page']).ad_group_id
 
+        if ad_group_id is None:
+            logger.error('Cannot handle url with no ad_group_id specified %s', entries[0]['Landing Page'])
+            return
+
         RawPostclickStats.objects.filter(datetime=dt, ad_group_id=ad_group_id).delete()
         RawGoalConversionStats.objects.filter(datetime=dt, ad_group_id=ad_group_id).delete()
 
@@ -221,6 +226,7 @@ bounced_visits=%s, pageviews=%s, duration=%s',
             landing_page = LandingPageUrl(entry['Landing Page'])
 
             assert landing_page.ad_group_id == ad_group_id
+            assert ad_group_id is not None
 
             source = resolve_source(landing_page.source_param)
             source_id = source.id if source is not None else None
